@@ -1,9 +1,9 @@
 [project]
 name = "times-doctor"
-version = "0.6.9"
+version = "0.7.0"
 description = "Diagnose and scan failed TIMES/Veda runs; optional LLM advice"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.12"
 license = {text = "MIT"}
 authors = [
   {name = "David Green", email = "david.green@csiro.au"}
@@ -18,9 +18,6 @@ classifiers = [
   "Operating System :: POSIX :: Linux",
   "Operating System :: MacOS",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Topic :: Scientific/Engineering",
   "Typing :: Typed",
@@ -32,7 +29,6 @@ dependencies = [
   "python-dotenv",
   "httpx",
   "psutil",
-  "tomli; python_version < '3.11'",
 ]
 
 [project.optional-dependencies]
@@ -68,7 +64,7 @@ url = "https://pypi.org/simple"
 
 [tool.ruff]
 line-length = 100
-target-version = "py39"
+target-version = "py312"
 
 [tool.ruff.lint]
 select = ["E", "F", "I", "UP", "B", "SIM", "S", "PERF"]
@@ -91,7 +87,7 @@ ignore = [
 "src/times_doctor/cli.py" = ["B008", "S603", "PERF203", "E741"]  # typer patterns, subprocess trusted, perf ok
 
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.12"
 warn_return_any = true
 warn_unused_ignores = true
 disallow_incomplete_defs = true
